Repository Initialisation
This commit is the initial commit of the repository. It contains the initial files and directories of the project.
The scaffolded Laravel project was created using the following command:
`composer create-project laravel/laravel .`

Included in the commit are the following files and directories to allow Docker to build and run the Laravel project:
- apache/000-default.conf
- Dockerfile
- start.sh
- supervisor/docker-worker.conf

The command used to run the project are contained in the README.md file.