Breeze Scaffolding Implemented
Implementation of Laravel Breeze scaffolding in Laravel. This is a simple, minimal, and easy to use authentication system for Laravel.
The Docker config needs updating to allow for the new files and directories created by the Breeze scaffolding.
This is a work in progress. The Docker config will be updated in the next commit via Pull Request.