Updated layouts
This layouts are designed to keep in line with the theme of the application. The navigation bar has been updated to include a link to the tasks page. The dashboard has been updated to include a link to the report page.
TODO: Format Todo list under view your tasks
Format Github Issues under Report An Issue